refactor(domain): turn output into an immutable value type
Merge `output_basis` into `output` as a single value type. Drop the
mutable `payment_address` cache and its upgrade-mutex; `addresses()`
now recomputes from `script_` on every call. Caller owns any caching
needed (same philosophy as the `header` and `script` refactors).

Net diff: -275 LOC after removing `output_basis.hpp/.cpp` and the
delegate boilerplate between the two classes.

The `validation_t::not_spent` sentinel kept its semantics but was
renamed (the inner struct + member could no longer share the name
`validation`); call sites under `blockchain/` and `database/` updated
accordingly. The C-API was regenerated to drop the stale references.
